+++ src/ConfigSymbols/PrintRemotes.d.ts
@@ -1,0 +1,3 @@
+declare const PrintRemotes: unique symbol
+
+export = PrintRemotes