+++ src/ConfigSymbols/PrintRemotes.lua
@@ -1,0 +1,2 @@
+local Symbol = require(script.Parent.Parent.Symbol)
+return Symbol.named("PrintRemotes")