+++ src/ConfigSymbols/ReceiveLogsFunction.d.ts
@@ -1,0 +1,3 @@
+declare const ReceiveLogsFunction: unique symbol
+
+export = ReceiveLogsFunction