+++ tests/client/ClientTest.client.lua
@@ -3,8 +3,8 @@ local ReplicatedStorage = game:GetService("ReplicatedStorage")
 local BridgeNet = require(ReplicatedStorage:WaitForChild("BridgeNet"))
 
 BridgeNet.Start({
-	send_default_rate = 60,
-	receive_default_rate = 60,
+	[BridgeNet.DefaultSend] = 60,
+	[BridgeNet.DefaultReceive] = 60,
 })
 
 local Object = BridgeNet.CreateBridge("Test")
@@ -14,9 +14,9 @@ local TestRemote = ReplicatedStorage:WaitForChild("TestRemote")
 TestRemote.OnClientEvent:Connect(function() end)
 
 Object:Connect(function(arg1, arg2, arg3)
-	print(arg1) -- Comment this out when stress testing. It'll lag heavily and produce incorrect results
+	--print(arg1) -- Comment this out when stress testing. It'll lag heavily and produce incorrect results
 end)
 
-while task.wait(1) do
+--[[while task.wait(1) do
 	Object:Fire("Hello", "world")
-end
+end]]